Dispose of tabs when they are cleared
This change changes the way tabs are disposed. Previously, their
disposal were tied to the project, which meant they weren't disposed
until the project was disposed. With this change, AppInspectionView
manages the disposal of the tabs it has created and disposes them when
the tool window is refreshed.

Now, we pass in a parentDisposable object in createTab, so inspectors
can chain their disposals to it. This change also updates existing
inspectors to use the new disposable object.

Lastly, this change adds a new field to keep track of the jobs related
to a "refresh" or "launch", and cancels them when the view tries to
launch new tabs.

Test: new app inspection view test.
Bug: 186225787
Change-Id: Ia5ab5793b437e3ac9c83d9f79d2817fb73ba4744

GitOrigin-RevId: 8bf14cfd78e864ac9879cd199ccc19539d3cc99e