Fix layout inspector infinite loading spinner
If you opened a project, ran the layout inspector, then closed and
reopened it, the next time you opened the layout inspector, it would
give you an infinite spinner.

This is because you normally need to wait for the agent to attach before
sending app inspection commands. However, when you attempt to attach the
agent after Studio has reopened, it finds the previous attach event.
Therefore, it immediately returns success, even thought it will still
take a few dozen more milliseconds before the device is actually ready.

In other words, this happens:

* Open a project the first time
* Open layout inspector
  - This requests that the app inspection agent attach
  - At this point, we poll all events we get from the device,
    blocking until we get an attachment confirmation
  - We launch layout inspector clients, which succeed
* Close the project
  - As a side effect, the agent on device is deattached
* Open the project a second time
* Open layout inspector
  - This requests that the app inspection agaent attach
  - At this point, we poll all events and immediately find the
    old one from the first attach event, reporting success
    without blocking
  - We attempt to launch layout inspector clients, which send commands
    expecting a response that never comes (as no agent is ready yet)
  - The inspector client init logic blocks forever
  - The agent eventually attaches, which is too late for the clients
    we tried to launch, but if you restart the app, things begin
    working at that point

Fixes: 184046769
Test: Couldn't figure out how to test, but at least it won't cause
      any regression
Change-Id: I48faa51a03d004fbd5c9373aa0697df41c5b645d

GitOrigin-RevId: b19fc5dc323acc17d3ceef3b630eaff325bb6008