Do not use reflection to get the baseline of view
This avoid the accessing of class, which improve the performance and
avoid the deadlock caused by acquiring the read/write lock.

The updated test cases use reflection to get the baseline, but the view
object in test environment is null so it doens't get the baseline. After
this CL, they uses the baseline gives by ComponentDescription.

Fixes: 174582760
Fixes: 174605585
Test: covered by UI test. Also updated related test cases.
Change-Id: I4f6aeb0c6f1c277649abcb749525f8d33d0156c4