Join Multi Line Logs At The Receiver
AndroidLogcatReceiver is in a great position to join multi line logs. Instead, it passes
this work down the line where it's harder to do.

This cl puts changes the way the entire stack handles logs. Logs are joined and parsed into a LogCatMessage
as early as possible, in the AndroidLogcatReceiver, and passed through the LogConsoleBase as a Json string
that's easy to convert back to a LogCatMessage where needed.

Ideally, LogConsoleBase would have been a generic class delaing with a T log entry type but since it is
hardwired to a string, this seems to be a reasonable option.

This cl leaves a lot of room for simplification by refactoring because a lof the classes and methods have
become very small as a result.

This refactoring will be done in subsequent cl's.

Fixes: 157171025
Test: Manually and existing coverage

Change-Id: Ib1e0badfd84aee143155a3b3d4887a0f6ad7f739

GitOrigin-RevId: e06c4f4bb23520133c5ad34e94334af1fbcfc73c