[Dsl] Extend external-to-model map tables
As well as the current four-element rows, consider five-element rows as
valid with the element at index 4 being a version constraint on the map
between surface syntax and model effect being valid.

This is rich enough to express the introduction of the `minSdk` property
as an alternate way of accessing the model previously accessed through
the `minSdkVersion` method.

Also make clear the nullability or otherwise of the fields in
ExternalToModelMap.

Bug: 184935026
Bug: 186623625
Test: existing (no change in functionality)
Change-Id: I6b13980592e9445e87cfe71e3da2f7c454eedeec

GitOrigin-RevId: eda2430124ff9adb552a48bb2b223b84b15c742e