[Dsl] Make Sdk properties version-aware
Use the new SdkOrPreviewTransform to model minSdk and targetSdk
properties.  Add tests verifying that "old" (pre-4.1.0) AGP behaviour is
unchanged, and adjust existing tests to reflect the new forms.

The operators for maxSdk need to be version-aware but do not need the
transform, as there is no overloading in the old-style maxSdkVersion
setter.

Fixes: 143196166
Bug: 148657110
Bug: 184935026
Bug: 186623625
Test: added, existing tests updated
Change-Id: Ida76c523546d908e2259272429a8603b452a8e97

GitOrigin-RevId: 007e6927e2acaf81077d373bca05e11a62d99056