Add deps on Kotlin 1.5 in android/BUILD
There are a couple tests which create sample Android projects using
the Kotlin version from the Kotlin IDE plugin---as opposed to the
version from TestUtils. So we need dependencies on Kotlin 1.5 too.

Eventually android:test_deps needs to be cleaned up. E.g. transitive
maven dependencies should not have to be listed manually.

Bug: 186464956
Test: AndroidLibraryTest.testRenameLibraryAndAddActivity
Test: CreateCppKotlinProjectTest
Change-Id: Idc57859cd208856ca5dd72cd3d3a9d6502c291d6

GitOrigin-RevId: 2947a1650f9ab31bbefd7700cab993eeb2f6696e