[AUA] Support for removing implementation properties
A number of properties of BaseExtension and some of its contained (in
the Dsl sense) classes are no longer available by exported interfaces.
This doesn't immediately matter for Groovy, though even the
implementation classes are expected to be separate by AGP 8.0.0, but it
has immediate ramifications for KotlinScript, where the build scripts
will not compile if they have references to implementation-class
properties that are not present (or write references to read-only
properties) in the interfaces.

Implement an upgrade RefactoringProcessor, with the slightly different
feature that its necessity is not characterized by a region, in that we
cannot from the Upgrade Assistant easily tell whether properties are
expressed in Groovy or KotlinScript build files, so we need to continue
to make this a mandatory upgrade if the region [current,new) intersects
a *region* of version space, rather than just a point.

This version restricts itself to removing only those properties
introduced by the 4.x templates for dynamic-feature modules, as a useful
staging point with a test case before the fuller implementation.

Bug: 186647828
Test: added
Change-Id: I5a07a3b63f15fb17ff73ce6d3283d49f2c9a9d89

GitOrigin-RevId: ae514ea43bcfdf884b1e05a1d0f6f2fb42b43215