IDEA-254379: Get rid of two kotlin plugins: replace KotlinPlugin lib with intellij.android.kotlin-dependencies module
KotlinPlugin is referencing jars that are created after compilation. That's why this library is not suitable for anything except being a RUNTIME dependency. Module, on the other hand, can be used in any scope.

intellij.android.kotlin-dependencies will be renamed to intellij.kotlin.plugin.community.main in the following commit.

GitOrigin-RevId: 168c597b6f385489c783c6731a9964b93c3821b8