Cause validateJdk to consider JDK per project
When a Gradle project is opened we validate if the JDK used for Gradle
is valid and if not we show an issue with some quickfixes. This check
validates the JDK returned by IdeSdks.getInstance().jdk, but after JDK
per project was introduced, this no longer represents the project JDK.

This change uses AndroidStudioGradleInstallationManager to find the JDK
used by the project when calling validateJdk.

Bug: 186868652
Test: JdkImportCheckTest
Change-Id: I4511ffa13e28f0bd32d3b8960ad9deeb6ccad834
(cherry picked from commit 7e57071df27aa0bfb51cfcb4c3800babf95b5d93)