Filter Live Literals element using compiler's metadata
If the compiler is providing metadata about the available literals, we
use that information to filter the LiteralsManager information.
This allows "hiding" literals that are not currently supported by the
compiler as "live" and notify out of date to the user correctly when
modifying those.

Bug: 186768446
Test: Existing tests pass, manually verified
Change-Id: I2ec2feb9675e18af39f1f61d3b6d3f52694689b6
(cherry picked from commit 737417fd2df68922c6468016dde74088569ad90b)