Add utility to control JFR
JFR is available in JDK 11, and can be enabled with the vm option
`-XX:+FlightRecorder`. JFR can be useful as a simple profiler.

This CL adds a bunch of internal actions to start, stop and dump
JFR recordings. The start action will enable JFR to record with
using the new configuration `prodprofile.jfc`. This configuration
is mostly a small subset of the default profiling configuration
that is available in the JDK: it disables most of the detailed
information, but keeps arounds stats in the following areas:
Java method profiling, GC info, file and socket tracing info.
The selected options should incur less than 1% overhead.

The plan is to have a small set of internal users to have this
enabled all the time and see how well it works out.

Bug: 187078308
Test: Start/Stop/Dump using internal actions
Change-Id: I6fbe988d83b76b05b4ae7240f3a46b61117caac0