Implement Android Test Retention UI in test config
It implements the "Enable Emulator Snapshots for test failures" checkbox
as well as other Retention configurations (max snapshots, compress
snapshots). The "enable" is a drop down that can be set to "Yes", "No",
"Use build.gradle settings". For the third option, it will grey out
everything in this session and just use whatever from build.gradle file.

The following items are to be implemented in follow-up commits:
 - Try pull values from build.gradle when set to "Use build.gradle
 settings"
 - Tooltips and helper links to documents.

Bug: 186904018
Test: GradleAndroidTestApplicationLaunchTaskTest, GradleConnectedAndroidTestInvokerTest

Change-Id: I2f32d790a60028d1ab88c3e165f0ddb0d7d51a16

GitOrigin-RevId: a127a75de576ff16c4412e813036befbc280c4ef