[AUA] Remove more implementation properties
The properties to be removed are derived from the 7.0 API removals
spreadsheet, which are themselves derived from a machine-generated text
file.  This text file is fine as far as it goes, but does not capture
some relevant semantics (such as renames to properties that happened
before the change in the interfaces).  For example, the rename from
aidlPackageWhiteList to aidlPackagedList is not currently modelled in
the Dsl layer, and so projects using the old name will currently not
have that name changed or removed (as appropriate to the kind of
module).

Bug: 186647828
Test: added
Change-Id: Icd9a2638771a8c8bf04290da1491e25f2a0e774c

GitOrigin-RevId: 5ae6033745e37bf468867950bba2a4ed615f61e6