Correct the color parsing function for Color class in Compose file.
If the argument order is not follow the function declartion by using
the named arguments, the color is wrong because we assume the argument
order is always rgba.

After this CL, the named argument is considered when reading the
arguments.

Also, when user set the color through the color picker, the replaced
expression will always be one of below formats:

- For Color(Int) or Color(Long), the format will always be
  Color(0xRRGGBBAA).

- For Color(Int, Int, Int) and Color(Int, Int, Int, Int) the format
  will be  Color(RR, GG, BB, AA)

- For Color(Float, Float, Float) and Color(Float, Float, Float, Float)
  the format will be Color(0.RRf, 0.GGf, 0.BBf, 0.AAf)

- For Color(Float, Float, Float, Float, ColorSpace), we don't support
  neither previewing nor updating color for it.

This CL also format the function call to make sure user pick up the
expected color from color picker:

- If any argument is given by a name, then all arguments will be named
  after picking a color.

- If the original function doesn't have alpha value, the alpha value
  will be assigned after picking a color.

- If any integer argument is represented in hexadecimal, all the
  arguments  will be represented by hexadecimal after picking a color.

Fixes: 186735797
Test: updated and covered by ComposeColorAnnotatorTest.
Test: manually.
Change-Id: I9b53362c4f7008a62a4f99a1a5f3f489e0e2fb50

GitOrigin-RevId: 105e26ad2ab06a3fbe5b5ff33ae6517b4045db73