Create a design tools plugin
This new plugin is contains:
- the designer module
- the nav-editor module
- the compose-designer module
- the customview module
- the layoutlib module
- the layoutlib jar
That corresponds to everything that needs direct access to layoutlib
classes.

It depends on the android plugin, and implements a new extension point,
layoutLibraryProvider, defined in LayoutLibraryLoader. That allows the
android plugin to still have access to a LayoutLibrary, while preventing
it from accessing the internals of layoutlib.

Bug: 187316741
Test: existing tests
Change-Id: I264d82e14cb206fa7e1443cb85a0e2bc1f79c200

GitOrigin-RevId: 4db1f6e44e6ccbd9d1d4d1bee3f4cf938655f339