[Dsl] Add dynamicFeatures as a version-aware property
The change to `val Mutable` in the interface came a long time ago,
apparently as far back as the 4.1 "new DSL" series.  There have been
various fixes to it since, apparently in response to problems with it in
KotlinScript, so I don't think it's worth chasing down exactly which
canary it first happened for; we can approximate with the release
version.

The past (assignment) surface syntax continues to work in Groovy, and
this is expected to continue for the forseeable future.  Both syntaxes
are acceptable for reading, but Groovy will continue to create new
elements using assignment.

Bug: 186624471
Test: added test for old syntax
Test: adjusted existing test
Change-Id: I05c8a16710e44ce0a28322192c35028547cbf925

GitOrigin-RevId: b66fc74bcddc38e105c5d834b488d358cb7cd0d2