[AUA] Ensure that there is a usage target
In the new interface to the Upgrade Assistant, users can (and in some
sense we hope they do) disable the classpath processor itself,
particularly in order to do component pre-upgrade / prerequisite steps
on their own.

However, this broke the implicit assumption that the classpath processor
is always on, which was used to populate the processor's target for the
preview window.  (Without a target, the preview window behaves more
erratically).

Workaround by unconditionally installing a fake Psi element as the
target for this refactoring.  This fake element doesn't have a
navigation target, but this is probably preferable to pointing at the
AGP version declaration when the user might have removed that from the
set of refactorings to execute.

Bug: 163020008
Test: added
Change-Id: I4fa2df1b162b53b52ca90a4d850bf1f26c0a9983
(cherry picked from commit eba5f35137d5a691109f05d0ff1144fab24acf5c)