[AUA] Fix threading update issue
If we are unlucky, we will refresh the processor associated with the
upgrade assistant twice in quick succession, and re-order the
reinitialization of the associated trees such that the one that is
discarded is used to initialize the nodes of the CheckboxTree.  This
then manifests as an upgrade that does nothing, as running the upgrade
first disables all the components of the processor, then re-enables
those that are ticked from the checkbox tree (which belong to an
otherwise-discarded processor).

Fix this by updating the processor field along with the tree, on the
dispatch thread.

Fixes: 186852662
Test: manual
Change-Id: Ie09be52aa7bd681b7f4caa412c1c0ee63d86abcc
(cherry picked from commit 7583e151b9f42b8391c14331cf2bc85b3cfdef7a)