Improve Log Line Batch Parsing
The motivation for this change is to be able to detect an EOM event in the log stream.

The problem is that the EOM is an empty line which can easily be embedded in a log message itself.

The solution used in this cl is to hold on the last entry in a batch of lines and post a delayed task that will flush it if no new batch arrives within a short timeframe.

To make the code more testable and readable, it's split into 2 methods:

One method parses the lines into an internal data structure, and the other handles the multi threaded aspect of the delayed task.

Most of the tests added in this cl use the first method to test the parsing code with a few tests that verify the delayed task is working correctly.

The existing test was changed in a couple of ways:
1. It was converted to Kotin, mostly in order to take advantage of String.replaceIndent() which makes multiline string constants much easier to read/write.
2. It changed the focus of the test from parsing individual entries to parsing entries that span batches. We already have good coverage for individual messages in the parser unit test.

Bug: 157171025
Test: Added
Change-Id: Ibe2ef75dca09e2b17270f6bf8098dc08fd5881a9