Recreate valid JDKs when importing configuration
When Android Studio is updated we ask users if they want to import
configuration from a previous AS installation. If they do so, we copy
the ProjectJDKTable and then recreate all their entries. This is needed
in order to update embedded JDKs that were overwritten by the update.

Jdk recreation was done using JavaSdk#createJdk(java.lang.String,
java.lang.String) but this method assumes that the path is a JRE and
thus does not generate all required class paths in all possible JDKs.

With this change we now use JavaSdk#createJdk(java.lang.String,
java.lang.String, boolean) that specifies if the created JDK is a JRE or
not and we keep only valid JDKs.

Bug: 187205058
Test: IdeSdksAndroidTest
Change-Id: I39744cdeb71a39ae3a23ad7c686908bbf2421714

GitOrigin-RevId: 6793133bd5adf03dd8c96358e176f0468011f47c