Recreate Gradle JDK when sync is run
... since there could be incorrect JDKs due to bug 187205058.

When we import a JDK from a previous Android Studio installation we also
import the JDK table and recreate the entries it has. Due to bug
187205058 some of those entries were created as JRE instead of JDK,
causing unresolved symbols issues.

This change causes the JDK entry used by the project to be recreated
when Gradle sync is run. If there are differences on the recreated JDK
and the one in the JDK table then it is updated, causing the class paths
to be updated as well.

Bug: 187205058
Test: JdkImportCheckTest
Test: IdeSdksTest
Change-Id: I84833f46ba30b76bd11fd19ae78cbdee37f906da

GitOrigin-RevId: 357ab7207c6ca452332d3baf7dfce1c587f8cc6e