Fix out of date notification
This fixes the notification that would should up after a new build when
this was the first build or happened after a clean.
This was caused by the ProjectBuildStatusManager taking a snapshot of
the file to be used to compare the file state. When the build happens
and the new set of "managed elements" is created, the snapshot would
only be updated if the previous state was Ready.

This also decouples the project build status from returned status,
allowing to track the out of date status of the file separately.

Bug: 187804630
Test: Manually verified.
Change-Id: I65432fe46ec04382e2704036ccd9cce84b07865c
(cherry picked from commit 70c8479c778661fa9a1eaf78e5fd24c2f434a468)