Fix the lambda link resolution.
We often saw an enclosing lambda many lines before being selected
instead of the lambda for the actual parameter.

To fix this: change the way we resolve lambdas by:
- always rely on line number information only if the lambda spans
  multiple lines.
- only use the lambda selector from the mangled class name when
  we find multiple lambdas on a single line.

Also in this change: select from the lambdas found for the parent
lambda immediate surrounding the lambdas found by line number.

If a lambda cannot be found: go to the beginning of the line
where the lambda was found via JVMTI.

The lambda selector is now expected not to contain anything but
the selector itself (without the starting "$" sign).

Fixes: 187969316
Test: Updated the test with the format expected from compose
Change-Id: I59cc11ce7cccc67b5c7e612bc17cdb507e5f0678
(cherry picked from commit 528c51673cc0b46577f2e9e9f715edb71db46d7f)