Persist layout in Compose Preview
This persist the layout and fixes a bug that would cause values not to
be correctly deserialized from disk (since the timing of events
differs).
This CL simplifies the order of events and reads the values correctly in
all scenarios.

Test: Existing tests pass
Fixes: 187967341
Change-Id: I83bdd89ea3ae590f8e468fbbf76c7a1781db47ed
(cherry picked from commit 5b50c65d3404a072469dc542acfeb676d9498e29)