Replace Request.listener with explicit listener parameter
`Request.myTaskListener` is a nullable and mutable field which is
initialized after construction.

This change replaces it with an explicit `@NotNull` parameter which
should allow making `GradleBuildInvoker.createBuildTaskListener` private
and moving its invocations back inside the `GradleBuildInvoker`.

Once done we will add the parameter back to the request object. However,
the default listener will now always be created allowing the field
to remain set to null.

Bug: 37797956
Test: n/a
Change-Id: I1a2a79b314d2c814b54693dbb9faa5ed4b343922

GitOrigin-RevId: 61e0f18e6c2d5edb048a208adf26ef05f417115f