[BA] Stop trying to download known plugins info
As we did not proceed with server provided file with plugins data
for now, implement a simpler version of a service that only parses
the bundled file without attempt to download file from the server.
Use this service as default for now.
Add test to verify local file parsed correctly.
Add test for download service for failed download.

Bug: 173784161
Test: added
Change-Id: I191f4affeb58c036dc65aaad8f344c18c7049509