Remove outdated Gradle "offline mode" check after build
`PostProjectBuildTasksExecutor` checks for an outdated error message
to detect potential issues with dependencies in the "offline-mode". In
fact, it doesn't receive error messages anymore and instead receives
just a "catch-all" exception messages "Cannot execute Gradle using...".

If this check is required it should be re-implemented as an build output
parser and a quick fix.

Having removed this we simplify handling of build results and make it
easier to return `ApkProvider`s directly from the `GradleBuildInvoker`.

Bug: 37797956
Test: n/a
Change-Id: I194a1befb7422a29af4475a3d8e274e39617242d

GitOrigin-RevId: c08742ce616fdc09ac4fc4ab06026bcd1ed0d9e3