[BA] Add explicit marker for non-compatible plugins
Before this CL the marker that plugin does not have a known version
compatible with CC was the absence of the corresponding version field in
the json plugin object.
It might be confusing for the future maintenance of the file.
Introduce the special value "N/A" that would mark the same. Internally it stays
the same for now and is parsed into a null.
Note: Surprisingly GradleVersion.parse accepts such strings and parses it into
a version with a value of 0.0.0. We need to be careful to not add
wrongly formatted versions as tests might not fail for them.

Bug: 173784161
Test: added
Change-Id: I9cbf4ace36a59028cb967a99428debf4ed6942f7