[Dsl] Fix some tests
It seems that buildToolsVersion has never been able to be set to an
integer, in Groovy or Kotlin -- at least not intentionally.  This means
that we can now implement a test for KotlinScript, given that we have
support for method calls.

Additionally, we have historically supported setting compileSdkVersion
in KotlinScript, but that is tricky from a type perspective, because
assignment only works for the declared type (String) as opposed to using
the overloaded setter.  Tweak the name mapping to discourage use of
assignment in KotlinScript, and adjust a test to match.

Bug: 148657110
Bug: 143196166
Fixes: 143196529
Test: existing test adjusted, .kts test added
Change-Id: I8b809ce9b55754e7d32d487cb3326d71a30d9750

GitOrigin-RevId: caa186b2192cd4300159bbf0f11aec54baa4d59f