Post build models should be fetched are required for assemble/bundle
In AGP 4.0 and older we rely on post build models to locate the
outputs. However, in many places where build is requested we locate
the outputs manually and it is rarely done correctly because of the
differences in how various versions of the AGP report locations.

This change is a step toward unified handling of build outpus via
`ApkProvider`s which will be returned by `assemble` and `bundle`
methods.

To simplify the code we will also request post build models when
build is requested by the user via make/rebuild actions. This step is
is fast and applies to old AGPs only.

Bug: 37797956
Test: n/a
Change-Id: Ib2bd905ccab472ca6c2b6e6d9df07564284c81a9

GitOrigin-RevId: ffe9ad3f01a645a0ae805ee0dc4901a753a6a64d