Introduce GradleBuildInvoker.executeAssembleTasks
instead of relying on callers to instantiate `BuildOutputAction`s which
are implementation details of `GradleApkProvider` and related classes.

This change also consolidates expansion of the assembled module list
in `OutputBuildActionUtil`. Previously we would expand the list with
feature modules in `OutputBuildActionUtil` and with base features in
`MakeBeforeRunTaskProvider`. Therefore, the `assembledModules` parameter
is now expected to get the list of modules the caller is interested in
(i.e. may need to request `ApkProvider`s). This will be better
documented in upcoming changes.

`GradleTaskRunnerFactoryTest` is deleted and not replaced with anything.
`ApkProviderOldAgpIntegrationTest` catches problems with incorrect
behavior of `ApkProvider`s and specifically `DYNAMIC_APP` tests
expansion of `assembledModules` list.

After this change all `GradleBuildInvoker` methods that produce APKs
will return `AssembleInvocationResult` which, in upcoming changes,
will be used to instantiate `ApkProvider` and `ApplicationIdProvider`.

Bug: 37797956
Test: ApkProviderOldAgpIntegrationTest
Change-Id: If5dc299630ae446b4f7b49761cac31dfb82d5d52

GitOrigin-RevId: ed3a9869eb315587bee1a489e61c3c69af31322c