Re-enable testLintJar on Windows
UrlClassLoader is optimizing for fast access to the underlying zip
files, locking them by default.

It seems safe to change the behavior of LintIdeClient and prevent it
from locking anything while running unit tests on Windows. Another
option would be to trigger a full garbage collection and remove the
soft references cached from UrlClassLoader, but it would rely on
implementation details.

Change-Id: I1ecfde835cd6d9ab21cee2ee5db85f6a6f4df6b6
Fixes: 188112031
Test: LintIdeTest.testLintJar

GitOrigin-RevId: cb7da74a70c06049b673781e52405f46996476c4