Do not create a search field unless it is needed
The implementation of SearchTextField.java would:
- create a AbstractPopup here: ComponentPopupBuilder.createPopup
- the popup would be registered as a child of the Application ???!
- the SearchTextField.removeNotify would close the popup

The component tree in the layout editor does not have a search field,
but AttachedToolWindow is creating a search field without ever using
it: thus the SearchField.removeNotify is never called the close the
popup. This causes a memory leak of the entire designer.

This fix is NOT to create a search field when not needed.

Fixes: 188820779
Test: N/A
Change-Id: I687ef2f07748e84d5d539715defb67885b71253f
(cherry picked from commit c6fbb8df3f4725f0223c3b824f4ead5705818916)