Attach agent only when it's necessary
We used to send an ATTACH_AGENT command every time we launch a target
and it worked for us because the transport daemon silently ignores all
commands except the first one that does the attaching. However, we
encountered issues like b/184046769. They were caused by the inherent
inability to associate ATTACH commands with responses.

This change makes it so that we check whether we need to send an
ATTACH_AGENT command in the first place, and only doing so if necessary.

Test: AppInspectionTargetTest
Fixes: 189229382
Change-Id: Idcab8571c81c4cecf8498577d4c80a030a62f29c

GitOrigin-RevId: d6007632923c1a83c79ca0dfc9b7e52b68b43d02