Do not register duplicated ResourceChangeListener in SceneManager
It is allowed to activate the SceneManager even it is activated.
When it happens, duplicated listener is registered to
ResourceNotifcationManager.
It may cause a memory leakage issue if the number of deactivation is
less than activation.

After this CL, the activation state depends on activate() and
deactivate() which is called last.

Bug: 188825161
Test: Add regression test.
Change-Id: I069de1fe0135b1489cfbceafb51570236310fa92
(cherry picked from commit fdeffd32576c18e074229c52c6e916155c47b415)