[DM] Fix some resizing issues
Don't show the column resizing cursor when mousing over the table
header, because we size the columns automatically so they can't be
resized manually.

Also change the empty text to use JBTable's StatusText, avoiding a
problem where the old empty label was too wide and forcing the table to
a minimum width. This made the first column appear to have extra space
even though the content was not that wide.

Bug: 188390099
Test: visual inspection
Change-Id: Iadb2ccb2cb6a63feb86d176a37c64ff5a5b4f272

GitOrigin-RevId: 4a823cbec5a096edacddd1b3a995f7b8bb1b41e5