Add migration messaging to relevant profilers view
This change replaces the network profiler monitor with a custom
migration panel that instructs the user to migrate to Network Inspector.
Similarly, it also replaces the System Events table with the same
messaging but to Background Task Inspector.

On the app inspection end, it introduces a new project service that
surfaces an API that controls the app inspection tool window, in this
case allows control over which tab should be focused.

All of the migration functionality are gated behind a studio flag.

Test: InstructionPanelTest and manual test.
Bug: 181294738
Change-Id: Ibced0da5c5b01acb790bdd738af2ba797318da4d

GitOrigin-RevId: 037c8a2a324d723f4714601726e6ce7fe037f48d