Do not create a new PhysicalDeviceTableModel
Do not create a new PhysicalDeviceTableModel in PhysicalDevicePanel
after receiving the devices from PhysicalDeviceAsyncSupplier

Having one instance of the table model will make it easier for
ActionsComponent to keep a reference to it. I'll need that reference to
implement the editing of device names.

The changes to PhysicalDeviceTable are fallout from that. The call to
setModel in the old version of PhysicalDevicePanel effectively fired a
table structure changed event and the new setDevices fires a table data
changed event. That triggers a new code path in two of the tests in
PhysicalDevicePanelTest that throws exceptions. Mock out the renderers
in those tests to avoid that.

Bug: 186786760
Test: PhysicalDevicePanelTest
Change-Id: Ic7045cbffda37eada32fceb27e40e160eb240d5d

GitOrigin-RevId: 103bb24a24f5dbd01b15fec50400bd3261b1e7e2