Take @Composable into account when locating a lambda
The compose kotlin plugin will use a custom mangling of the lambda class
when the lambda parameter is a @Composable with no capture scope.

On the Studio side we are able to detect when a lambda parameter is
annotated with `Composable` but it is deemed too hard to detect when
the compiler needs to create a scoped variable.

The fix is to detect when a @Composable lambda parameter is found in
the code. Such a lambda may be mangled with a custom name, and if it is
the siblings of the lambda can no longer be found by looking at the
class name of the lambda.

Also if the wanted lambda actually has a custom mangled name we can no
longer trust the nesting level.

In cases we we no longer can identify an exact lambda expression we
will mark the start line found by JVMTI instead of the exact lambda.

Fixes: 189880539
Test: Added unit tests
Change-Id: I18e271e79029c4189935c3041a243fbc09338446

GitOrigin-RevId: 0834ce16adb6754ec589946fa1d7ca8dc1819390