[AUA] Disable tree on loading and running states
The cause of the bug is that tree stays enabled during
run or loading process allowing users to browse through
the steps making details appear under the loading label
on smaller screens. This  is not an intended behaviour.
To fix clear the tree selection on loading state to hide
the details and disable the tree to disallow new step
selections.

Bug: 188496458
Test: N/A
Change-Id: I94a8699bfbc43219577708ffb13d67587648f579

GitOrigin-RevId: 7e5424ce92e40292e2e887dc55876df4094ef0ad