Test for JDK re creation during sync
Having corrupted JDK information on the Project JDK table is can lead to
missing symbols and other issues. In order to mitigate this, we re
create the JDK entry in the JDK table if there are differences during
sync. We currently have unit tests to verify the needed parts for this
work as expected but we lack any integration test to check that syncing
indeed causes the JDK table to be recreated. This changes adds a couple
of tests to cover it.

Bug: 187205058
Test: JdkRecreationIntegrationTest and IdeSdksAndroidTest
Change-Id: Ibf842fa093da9dd40e33a48bbb85dce802671dde

GitOrigin-RevId: 64a88e23267e2845939f6b1f4740694c108584cc