[MPSS] Re-route facet methods to correct modules
This change modifies the getInstance methods for AndroidFacet, JavaFacet
and GradleFacet to use the module within the LinkedAndroidModuleGroup if
it exists. This will be set when module per source set is enabled.

This will ensure that the JavaFacet and GradleFacet are pulled from the
holder module and the AndroidFacet is obtained from the main source set
module.

This change currently does not affect behaviour as module per source set
is disabled.

Test: will be added in a following change
Bug: N/A
Change-Id: Iee8112252b281afb4af731ce128a5a47f1f40e61

GitOrigin-RevId: 3f62bb008ba3e2717e58516a0aa5f49b92d99d42