Stop using dx.jar for expression evaluation
When evaluating a kotlin expression that generates multiple classes
(lambdas) they cannot be interpreted on the host. So they are dexed
and sent to device.

To do this, the code was being built against a dx.jar,
but not shipped. Then it looked for it in the SDK and dynamically
reloaded the class that depended on it.

This change removes all of that and uses instead the d8 bundled
with Studio which already fixes some expressions that returned
incorrect values.

Test: N/A
Bug: N/A
Change-Id: Ib20b60f4e0cd6a4ff87cc066f2ca5019ef5d4098

GitOrigin-RevId: 70cfb0eebec3166771145ff08d3b83627ac7bbb7