Adjust LatestKnownPluginVersionProvider
When running under bazel, the offline maven repository can be placed at
an path relative to the (possibly absolute) temporary directory found
from `java.io.tmpdir` -- in particular, this is how it is set up in
presubmit.

This location was completely unknown to
LatestKnownPluginVersionProvider, which led to confusing results, as the
-dev version of AGP would typically be in the offline repository, but
other versions might be in other locations actually found by the
LKPVProvider, leading to spurious return of old versions of "latest".

Bug: 148081564
Test: existing test adjusted
Change-Id: Ic432ea0034c1ab451066064019042783acf81784

GitOrigin-RevId: 0b1c794136b8423ff73110e328d4243f3a08c913