Partially fix support for true APK file lists
Historically build was expected to produce an APK file or a collection
of APK files in one directory and a single `File` object was enough to
describe such output. The file object would either point to the APK or
to the directory containing multiple APKs. This is changing and we need
to extend the support for exact lists of APKs to all consumers.

In this and upcoming changes we gradually replace APIs returning a
single `File` object with APIs that return a collection of `File`
objects and hide the the handling of different AGP versions behind these
APIs. Ultimately, the location of APK files should only be requested
through `ApkProvider` objects.

Bug: n/a
Test: existing (ApkProviderIntegrationTestCase)
Change-Id: I3206ef4bd17abd777d97c7a41ee2d3e96213af1c

GitOrigin-RevId: 027ba8bab99a4e66109846b7116c0acdeb47e158