[Dsl] Remove argument to addGoogleMavenRepository
All versions of Gradle that are compatible with versions of AGP
supported by Studio have the built-in `google()` shorthand for the
Google maven repository.

Removing this argument from the interface function is, of course, a
breaking change, but this is the last vestige of dependency on the
rest of the android plugin.

Bug: 190504109
Test: existing tests adjusted
Change-Id: I86f8ba040856ad513567b4579351c4967f2db5d3

GitOrigin-RevId: a5da0138d05799ddf2e2226c1b7e9134297991c7