[AUA] Refresh projectBuildModel consistently
We need to refresh the projectBuildModel in case the user has made
changes to files between successive calls to findUsages in the same
processor (for example, if displaying a preview/usage view and the user
has made some manual edits).  However, we do not want to refresh it more
than once, as it is potentially expensive: but individual
AgpComponentRefactoringProcessors are used elsewhere than in the
upgrade.

Deal with this by differentiating between constructing a processor as a
child of an AgpUpgradeRefactoringProcessor and constructing one ab
initio, and only refresh in the component if it does not have a parent.

Bug: n/a
Test: existing
Change-Id: I7c62a49f8bdf582df73bc5ce31a677f3df177439