Database Inspector: fixes an issue with db cloning
`sqlite3` clone operation does not cope well with special tables used by
the Room library.

This change replaces relying on the clone operation with performing a
WAL checkpoint and copying/moving the main DB file.

As a side-effect, this is likely going to improve performance as in most
cases, we're just going to move the DB file instead of making a copy.

Bug: 161081452
Bug: 191010086
Test: Reproducing the Room special case non-trivial. Base behaviour
      covered by existing tests.

Change-Id: Iac2eeedfe6d11fe6bea981128d80cafbf00a8966

GitOrigin-RevId: 5cd2d86f1a58fadb3780fca8d8d3650d7a6a0ce3