Delegate exception in multi-inspector tabs
This change delegates the handling of inspector exceptions in
multi-inspector tabs to the tabs themselves. When an inspector goes
down, the tabs can use the provided ideServices to show a toast, and/or
change their UI to reflect the state of their app. When all inspectors
go down, AppInspectionView will replace the inspector tab with an empty
state message.

Exceptions generated by inspectors in single inspector tabs (those that
extend SingleAppInspectorTabProvider) are handled as usual by App
Inspection.

Note, in a follow-up change, support for restarting individual inspector
tabs will be added.

Test: Updated existing tests.
Fixes: 187421789
Change-Id: Icb57521544f478ba7984c496f5ffd61b57c70b31

GitOrigin-RevId: af455a46f98d8ded44a1f91f90ed57c103c2a89a