Fixing metrics for atf
Our usage data seems bit off. Trying to fix in 2 places:
1) Unnecessary dependency to render metrics. If render metric somehow
does not return the right value atf usage metric is skipped. This is not
ideal for learning user behaviour
2) Too large data - Based on queries, there're alot of verbose or info
issues that is included. For now report only what the user sees so that
usage metric is not affected by this.

Bug: 191303537
Test: Manually tested
Change-Id: I8d6be1f76b3a1665dedd1b7971d1cf3b965fdf2e

GitOrigin-RevId: 9b41528fb2b556118628a23bd87260882a0ff8aa