[AUA] AGP 7.0.0 requires Gradle 7.0.2
This will have automatically been handled in Arctic Fox builds of
Studio.  The main development branch, though, needs to know about the
requirements of the other development branches in order to be able to
properly compute the upgrade to them.

The properties of third-party plugins are assumed not significantly
changed by this (the divergence between Gradle 7.0 and 7.0.2 should be
minor).  Update a comment relating to safeargs to reflect the lack of
removal of a deprecated property in AGP 7.0.0.

Bug: n/a
Test: existing test extended
Change-Id: I88a94b0e7f0256f57b8764cb3dc29c7155455b83

GitOrigin-RevId: 5b11455de3ddb3c42384a82fd22b997b05a4e61e