[safeArgs] Suggest companion object imports
In addition to classes, we are able to suggest to import
those members of the companion objects.

While implementation, some issues are exposed and thus
addressed in this cl, like follows,
1) There's a chance, for a given unresolved function, there
might be multiple import suggestions, and they might not
be differentiated, like this,
https://screenshot.googleplex.com/6u2DfLiUzAaLpGo.

Thus, we explicitly include extra "container" identifier
info, like this,
https://screenshot.googleplex.com/9PgmThodsrCCJVG.

Basically, for a given function, we include the corresponding
containing fqcn for the later displaying/differentiating purposes.

2) Instead of the method icon, function icon is more appropriate,
and thus it's being used.

Last, b/191785103 is filed - it's unfortunate there might be 2 sets
of imports to offer, and this is very confusing. We might need to
find a way to consolidate.

Bug: 190853314
Test: updated
Change-Id: I17b54df9df79024b8e5f12380fb1ad8093cc90f4

GitOrigin-RevId: fbb53cfa6cfca20d81ecfb92ee4038fd9388af58