[AUA] Track file-complexity of upgrade
The number of build files is not quite the same as the number of
modules, not least because the settings file counts, but also because of
potential `apply from:` or similar including constructs.  However, it's
probably a reasonable correlation of the module-complexity.

Also adjust the sync listener so that it closes over the
known-at-instantiation-time values of the number of usages and number of
build files.

Fixes: 185154780
Test: existing tests adjusted
Change-Id: Iae748e2076f9cab68cd8fc16accee513f336202c

GitOrigin-RevId: 20f9906b1b461057cd85503870590d942242286e