[Dsl] Support plugins with version and no explicit apply
The `apply` can be implicit, defaulting to false in settings and true in
project build files.  House style might well be to use the implicit
`apply` at least in the settings file, so support adding it, preferably
with a two-argument form, which itself uses the newly-nullable third
argument to applyPlugin().

Fixes: 191625008
Test: added, existing tests adjusted
Change-Id: Iea3041b8954ffc378af396cb14925a4338dcbee5

GitOrigin-RevId: d9b3b4aa60edb293eb299427c09865788e40fea8