[Wear Pairing Assistant] Implement reverse forwarding
Instead of creating a port forward between
the Phone and WearOS Emulator Host on port 5601,
create it on port 5602, and then reverse forward
the WearOS Emulator port 5601 to this new
Host port (5602).
This CL does not update the UI to allow
multiple Phone-WearOS connections. It just changes
the way the current single connection is done.

Bug: 191732929
Test: Current tests, just a different way to connect
Change-Id: Ic9fd3ea83d97b632e92b725c67c5fd6bd4ae9071

GitOrigin-RevId: 9b8db31f3532063e3f5c400edc9819c2596550eb