Move AGP based tests in __other target to agp-integration-tests module
This reduces the number of total shards from 10 to 9. Test time across
the affected targets remains about the same on linux. On windows, the
test time is reduced by 2-3 minutes.

TestRClassesTest and GeneratedResourcesTest are split out of
LightClassesTest because they subclass AndroidGradleTestCase, and
need to be in separate files in order to be moved.


Bug: 190234522
Test: NA
Change-Id: If294616016420f222ba1e10d5dfae3074f10ec37