Update package references after ui-tooling migration
After the migration that split the ui-tooling into two artifacts (one with
the same name and ui-tooling-preview), some classes were moved to a
different namespace, for example ComposeViewAdapter. Since use the FQN of
such classes in Android Studio, we need to update these references to be
prepared when the Compose release candidate artifacts are released.
This CL detects the right namespace to use.

This CL also removes the old androidx.ui. namespace references.

Test: updated existing tests
Bug: 191760028
Change-Id: Ic2e1bfa8d1da9ca3c33c3481bf744e7fdecfcd7d

GitOrigin-RevId: e2a7ac5f0974c9958212dc7861d3376ad180940d