[Dsl,.kts] Adjust lintOptions.textOutput encoding
Previously, when the Dsl interface was expressed in Java, textOutput was
an anomaly among lintOptions properties because of the additional String
setter methods, meaning that there was no automatically-generated
writeable property in the KotlinScript Dsl.  Now that the interface is
expressed in KotlinScript, the writeable property is available.

This causes a change in the emitted Dsl when writing textOutput, which
would theoretically be a problem with older (pre-4.1) versions of AGP.
However, in practice nothing actually writes lintOptions.textOutput, and
the planned use in the Upgrade Assistant (which will write a
textOutput-like property) will target AGP versions around version 8.0.

Bug: 144403889
Bug: 191643297
Test: existing tests adjusted
Change-Id: I7a764744f355c9c0ab9443813b21f1c5444bb6ba

GitOrigin-RevId: 6c27b6dd1b9f6936693d0ded529e00adc25167a8