Add API to incrementally build a file and its dependencies
The ProjectSystemBuildManager only exposes an API for compiling the
entire project. For certain cases (rendering, for example) requires
running the compile action for a specific file. Currently, all such
usages directly call into GradleBuildInvoker with the module containing
the file.

This CL adds a `compileFileAndDependencies` method to
ProjectSystemBuildManager which can be used to incrementally compile a
given file and all its dependencies, and adds implementation to
subclasses of ProjectSystemBuildManager.

Bug: 184199691
Test: Existing Gradle tests pass
Change-Id: Idb59160c2a8a5a1087c757efcd1298f22b8f499a