Use incremental file build API to issue compose builds
For non-gradle projects, Compose preview currently calls
ProjectSystemBuildManager.compileProject which triggers a full project
compilation for each refresh.

This CL changes the call to
ProjectSystemBuildManager.compileFileAndDependencies which allows
incremental compilation for a specific file and its dependencies. This
is what Compose Preview does with incremental Gradle builds for the
module being previewed and all its dependencies.

Bug: 184199691
Test: Existing Gradle tests pass
Change-Id: I782d8eede8cfd96be48a3e468a17f0a539ec5743

GitOrigin-RevId: 7a0894373226e2cba1d730d483398096e5644e04