[Dsl] Support lint as a somewhat-modified lintOptions
Now that the Dsl interfaces Lint and LintOptions are distinct, the our
Dsl representations are likewise more decoupled.  We no longer need to
support check(), nor do we have to have any version constraints.  We
also no longer support old-style helper methods in KotlinScript, as they
are not part of the interface.

Bug: 191643297
Test: added
Change-Id: I0fa49137aa5641c971b9ee77a876985fcbc38166

GitOrigin-RevId: 998c1791ebca57110546e08980240cca85fd2ffb