Fix RadioButton completion
And make the lambda addition more restrictive. Curly brackets were being
added at the end of any completion as long as it contained a functional
type as a parameter as the last required parameter, so:

fun AMethod(a: String, b: () -> Unit, c: Int = 3)

For AMethod${caret}, this could add the lambda at the end as:
AMethod(a = ) {
}

This behaviour does not match the trailing lambda behaviour of intellij
and this CL restricts it so there can not be additional parameters after
it.
Also, since the main use case is onClick, this CL restricts the
completion to lambdas with no parameters.

Fixes: 191977716
Test: Added regression testing, other tests pass
Change-Id: I6efa75a90e48b6d836709e0256848333beb6908d

GitOrigin-RevId: 58dc928aeff3791beb2567ae786b71ef77e6f260