Step to variant specific GradleIdProvider's
This change may result in `ApkProvisionException`s being thrown where
previously an incorrect value was returned. This should not surprise
callers since they already handle it and should not happen in general
except the UI code that displays run configuration statuses and handles
the exception by displaying it in action tooltips.

Bug: n/a
Test: n/a
Change-Id: I7ad0b59cde9932bbd43590de31af3e182eff8d16