Fix making composite builds
and add a test case to `ApkProviderIntegrationTestCase` to prevent
regression. Gradle roots for modules from included builds was not
detected correctly. In the context of `GradleBuildInvokerImpl` it is
supposed to be the included Gradle project root rather than the root
of the Gradle project attached to the IDE project.

This change also adjusts `projects/compositeBuild` test project to make
apps from the included builds buildable.

Bug: n/a
Test: ApkProviderIntegrationTestCase
Change-Id: Ib0c9bca10f5ce4552f605f4395649641d38ad47e

GitOrigin-RevId: 0db6c0019243db37c8aac7e1a722e6dde8434823