Fix testLoadingPane
AppInspectionInspectorRule installed a default command handler that
would send a layout event upon connection. In testLoadingPane this
caused a race condition where sometimes this event would cause the
loading pane to already be hidden by the time we started waiting for
the loading pane to show. Added a mechanism to prevent the default event
from being sent.

During investigation I also found a problem where ApplicationRule wasn't
cleaning up correctly in the case where the initial application was
null, so fixed that as well.

Fixes: 183942017
Test: this is a test fix
Change-Id: I3e0c621e8ca86e6bcd836ae97ac494fa65f9c5fd

GitOrigin-RevId: 6032ff25e79e107fa587c46f87f1135b175c8ee4