Adjust the android View move logic
Before this change we moved the AndroidViewsHandler from the
view hierarchy to the compose node where the view was embedded by
the compose application.

However the AndroidViewsHandler was a shared view for all the
embedded views in a compose app.

Fix this by explicitly finding the AndroidViewsHandler and move
its children to the appropriate compose node. We can ignore the
AndroidViewsHandler if all children were moved.

Note: this fix requires a similar change to the compose inspector.

Fixes: 191695903
Test: Adjusted an existing test to verify the new behaviour.
Change-Id: Ic8240cdbb0649a73ba892dc71b496d9af82fe69b