Fix CreateNewMobileProjectTest flakeyness
CreateNewMobileProjectTest was flakey due to unknown issues in
Swing JRootPane creation. Occasionally it will be created with
screen coordinates that do not match getLocationOnScreen. This
causes the robot to not be able to correctly locate the button
and thus missing the intended target.

This workaround just avoids using the mouse when we can detect
that the location of the pane is "incorrect", and in such case
use the keyboard to press the enter key instead of clicking.

//tools/adt/idea/android-uitests:CreateNewMobileProjectTest              PASSED in 371.7s
  Stats over 1000 runs: max = 371.7s, min = 162.9s, avg = 232.2s, dev = 38.0s

Bug: 181614727
Test: CreateNewMobileProjectTest
Change-Id: Id2cadaab9e5c5affc1700c751e819e3552c77eeb