[Dsl] Bandage a cast exception / assertion failure
When opening the PSD on some sample projects, the structure leads
us to create a reference to an element whose name (in the sense used
by the PSD) does not identify that element, making us construct a
RawText Dsl element instead.  This is not expected by the
implementation of FakeArtifactElement, leading to assertion failures
in -dev builds and a cast exception if the assertion is compiled out.

This fix removes the PSD exceptions but does not fully with the
subsequent problems in the PSD's functionality.

Bug: 174548560
Test: manual
Change-Id: I9ef02db578311ffbc4b2bc6810935589ce3afc4d

GitOrigin-RevId: 06a5373043ced0558c7c7e48df2ebdd5878f9e21