[Dsl,PSD] Support PSD's use of ReferenceTo
For operations within the Dsl module, ReferenceTo objects need not
retain a reference to their scope, as they are externalised by the
backends within that scope, so the external text generated is correct
for whatever scope is relevant.

For some other tools, including the PSD, the ReferenceTo object is
queried for a string representation independent of its context, and
indeed that string representation must round-trip into an equivalent
ReferenceTo using createReferenceFromText() -- as in dslText() and
setDslText() extension methods on ResolvedPropertyModel.  This means
that the ReferenceTo itself needs to have some idea of its context or
scope as well as the element it is referring to.

This also allows us to remove most of the special-casing of
SigningConfig models in ReferenceTo, by choosing an appropriate context.

(This also slightly reduces the confusion about whether the name
produced by toString() and used in the PSD is an internal name, an
external name, or indeed some third option: toString() produces
something which is more like an internal name, using dotted notation to
dereference maps rather than whatever the external language expects.
There is still plenty of confusion possible in this area.)

Fixes: 174548560
Test: existing, added
Change-Id: I17d4304b048dcf77d3650845f497fa4ed1214f1d